Patterns: split up the block editor inserter patterns tab into separate component files  (#55315)
