+++ packages/block-editor/src/components/inserter/block-patterns-explorer/index.js
@@ -8,9 +8,9 @@ import { __ } from '@wordpress/i18n';
 /**
  * Internal dependencies
  */
-import PatternExplorerSidebar from './sidebar';
-import PatternList from './patterns-list';
-import { usePatternsCategories } from '../block-patterns-tab';
+import PatternExplorerSidebar from './pattern-explorer-sidebar';
+import PatternList from './pattern-list';
+import { usePatternCategories } from '../block-patterns-tab/use-pattern-categories';
 
 function PatternsExplorer( { initialCategory, rootClientId } ) {
 	const [ searchValue, setSearchValue ] = useState( '' );
@@ -20,7 +20,7 @@ function PatternsExplorer( { initialCategory, rootClientId } ) {
 		initialCategory?.name
 	);
 
-	const patternCategories = usePatternsCategories(
+	const patternCategories = usePatternCategories(
 		rootClientId,
 		patternSourceFilter
 	);
