+++ packages/block-editor/src/components/inserter/block-patterns-explorer/pattern-explorer-sidebar.js
