+++ packages/block-editor/src/components/inserter/block-patterns-tab/pattern-category-preview-panel.js
@@ -1,0 +1,48 @@
+/**
+ * WordPress dependencies
+ */
+import { useRef, useEffect } from '@wordpress/element';
+
+import { focus } from '@wordpress/dom';
+
+/**
+ * Internal dependencies
+ */
+
+import { PatternCategoryPreviews } from './pattern-category-previews';
+
+export function PatternCategoryPreviewPanel( {
+	rootClientId,
+	onInsert,
+	onHover,
+	category,
+	showTitlesAsTooltip,
+	patternFilter,
+} ) {
+	const container = useRef();
+
+	useEffect( () => {
+		const timeout = setTimeout( () => {
+			const [ firstTabbable ] = focus.tabbable.find( container.current );
+			firstTabbable?.focus();
+		} );
+		return () => clearTimeout( timeout );
+	}, [ category ] );
+
+	return (
+		<div
+			ref={ container }
+			className="block-editor-inserter__patterns-category-dialog"
+		>
+			<PatternCategoryPreviews
+				key={ category.name }
+				rootClientId={ rootClientId }
+				onInsert={ onInsert }
+				onHover={ onHover }
+				category={ category }
+				showTitlesAsTooltip={ showTitlesAsTooltip }
+				patternFilter={ patternFilter }
+			/>
+		</div>
+	);
+}
